requirements.txt
aiohttp==3.8.3
discord.py==2.2.3
googletrans==3.1.0a0
psutil==5.9.4
python-dotenv==1.0.0
discord-pretty-help==2.0.2
